+++ tests/src/rules/prefer-default-export.js
@@ -170,26 +170,30 @@ context('TypeScript', function () {
         // Exporting types
         semver.satisfies(tsEslintVersion, '>= 22') ? test({
           code: `
-          export type foo = string;
-          export type bar = number;`,
+            export type foo = string;
+            export type bar = number;
+            /* ${parser.replace(process.cwd(), '$$PWD')} */
+          `,
           ...parserConfig,
         }) : [],
         test({
           code: `
-          export type foo = string;
-          export type bar = number;`,
+            export type foo = string;
+            export type bar = number;
+            /* ${parser.replace(process.cwd(), '$$PWD')} */
+          `,
           ...parserConfig,
         }),
         semver.satisfies(tsEslintVersion, '>= 22') ? test({
-          code: 'export type foo = string',
+          code: 'export type foo = string /* ' + parser.replace(process.cwd(), '$$PWD') + '*/',
+          ...parserConfig,
+        }) : [],
+        semver.satisfies(tsEslintVersion, '> 20') ? test({
+          code: 'export interface foo { bar: string; } /* ' + parser.replace(process.cwd(), '$$PWD') + '*/',
           ...parserConfig,
         }) : [],
         test({
-          code: 'export interface foo { bar: string; }',
-          ...parserConfig,
-        }),
-        test({
-          code: 'export interface foo { bar: string; }; export function goo() {}',
+          code: 'export interface foo { bar: string; }; export function goo() {} /* ' + parser.replace(process.cwd(), '$$PWD') + '*/',
           ...parserConfig,
         }),
       ),